+++ src/main/java/io/zoemeow/dutapi/objects/NewsType.java
@@ -1,6 +1,6 @@
 package io.zoemeow.dutapi.objects;
 
 public enum NewsType {
-    General,
+    Global,
     Subject
 }